2nd

+++ TicketApp.WebApi/Models/Order.cs
@@ -3,10 +3,10 @@
     public class Order
 
     {
-        public int Id { get; set; }
-        public int TicketId { get; set; }
+        public long Id { get; set; }
+        public long TicketId { get; set; }
         public virtual Ticket? Ticket { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public virtual User? User { get; set; }
         public DateTime Date { get; set; }
         public bool IsPaid { get; set; }
